Move traversal to its own module
